0415pjt

+++ 05_django_model_relation_l/articles/views.py
@@ -96,5 +96,6 @@ def comments_create(request, pk):
 def comments_delete(request, article_pk, comment_pk):
     if request.user.is_authenticated:
         comment= get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
+        if request.user == comment.user:
+            comment.delete()
     return redirect('articles:detail', article_pk)
